Translate choices and use an empty choice as default
By default types have the default translation domain in this bundle,
which is good, because I want to translate the labels in user-land, but
the boolean type sets a different domain, so my label will be
translated in that domain instead. This fixes the issue and only
translates the choices in the bundles domain.