Remove deprecation notice in PHP 8.1
Adds a `TreeBuilder` return type to `Lexik\Bundle\FormFilterBundle\DependencyInjection\Configuration::getConfigTreeBuilder()` to remove a deprecation notice triggered by the missing return type in PHP 8.1. That does not conflict with older PHP versions, since at least PHP 7.1 is required in composer.json for the bundle.

Also tidies up the "BC layer for symfony/config 4.1 and older" from that method as in composer.json at least Symfony 4.4 is required for the bundle and the BC layer is not needed anymore.